nobahar/backend/requirements.txt
Django==3.2
djangorestframework==3.13.1
markdown
django-filter
python-dotenv==0.19.2
psycopg2-binary==2.9.3
djangorestframework-simplejwt==5.1.0